+++ Sources/KeyboardKit/Callouts/View+SecondaryInputCallout.swift
@@ -14,13 +14,10 @@ public extension View {
      This modifier can be applied to any view that should be
      able to present a secondary input callout.
      */
-    func secondaryInputCallout(
-        for keyboardContext: KeyboardContext,
-        context: SecondaryInputCalloutContext = .shared,
-        style: SecondaryInputCalloutStyle = .standard) -> some View {
+    func secondaryInputCallout(style: SecondaryInputCalloutStyle = .standard) -> some View {
         return ZStack {
             self
-            SecondaryInputCallout(context: context, style: style)
+            SecondaryInputCallout(style: style)
         }.coordinateSpace(name: SecondaryInputCalloutContext.coordinateSpace)
     }
 }