+++ Sources/KeyboardKit/Emojis/EmojiKeyboard.swift
@@ -73,9 +73,10 @@ public struct EmojiKeyboard: View {
 @available(iOS 14.0, *)
 struct EmojiKeyboard_Previews: PreviewProvider {
     static var previews: some View {
-        SecondaryInputCalloutContext.shared = SecondaryInputCalloutContext(context: FakeKeyboardContext(), actionProvider: StandardSecondaryCalloutActionProvider(), actionHandler: FakeKeyboardActionHandler())
-        return ScrollView(.horizontal) {
+        ScrollView(.horizontal) {
             EmojiKeyboard(emojis: Array(Emoji.all.prefix(50)))
-        }.environmentObject(ObservableKeyboardContext.preview)
+        }
+        .environmentObject(ObservableKeyboardContext.preview)
+        .environmentObject(SecondaryInputCalloutContext.preview)
     }
 }