+++ Sources/KeyboardKit/System/SystemKeyboardButton.swift
@@ -33,11 +33,16 @@ public struct SystemKeyboardButton: View {
     
     @EnvironmentObject private var context: ObservableKeyboardContext
     @EnvironmentObject private var inputCalloutContext: InputCalloutContext
+    @EnvironmentObject private var secondaryInputCalloutContext: SecondaryInputCalloutContext
     
     @ViewBuilder
     public var body: some View {
         SystemKeyboardButtonContent(action: action, text: text, image: image)
             .standardButtonStyle(for: action, context: context)
-            .keyboardGestures(for: action, actionHandler: actionHandler, inputCalloutContext: inputCalloutContext)
+            .keyboardGestures(
+                for: action,
+                actionHandler: actionHandler,
+                inputCalloutContext: inputCalloutContext,
+                secondaryInputCalloutContext: secondaryInputCalloutContext)
     }
 }